backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
h11==0.16.0
idna==3.10
mysqlclient==2.2.7
outcome==1.3.0.post0
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.5.0
PySocks==1.7.1
python-decouple==3.8
pytz==2025.2
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
